requirements.txt
# LocaLocaLocalize Dependencies

# Browser automation
playwright==1.35.0

# OCR and image processing
easyocr==1.7.0
langdetect==1.0.9
pillow==10.0.0
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3

# Google Cloud Vision API
google-cloud-vision==3.4.3
requests==2.31.0

# Data handling
pandas==2.2.0

# Configuration and file handling
pyyaml==6.0
python-dotenv==1.0.0

# Date/Time Parsing
python-dateutil==2.8.2
regex==2023.12.25

# Utilities
tqdm==4.65.0
colorama==0.4.6 